+++ CHANGELOG.md
@@ -7,7 +7,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Next
 * Update nps-survey element: logic and mobile layout
-
+* Fix the bootstrap date icon visualization
 
 ## [1.4.9] - 05.02.2025
 * Add new "list/list item" element where it's possible to sort the list with checkbox via drag and drop or keyboard
@@ -20,7 +20,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Fix error in the label and tooltip visualization
 * Fix telephone number form field validation
 * Columns displayed in grid not flex
-
 
 ## [1.4.8] - 11.10.2024
 * export text element with .editor() function where htmlEditorConfig can be customized